+++ src/app/store/reducers.ts
@@ -4,7 +4,7 @@ import { AccountsState, accounts } from '../accounts/reducer';
 import { InventoryState, inventory } from '../inventory/reducer';
 import { ShellState, shell } from '../shell/reducer';
 import { ReviewsState, reviews } from '../item-review/reducer';
-import { Settings } from '../settings/settings';
+import { Settings } from '../settings/reducer';
 
 // See https://github.com/piotrwitek/react-redux-typescript-guide#redux
